feat: user api 연동 (#31)
* fix: 게시물 등록 오류 수정

서버 값이 camelCase로 변경되어 이에 맞게 수정했습니다.

* feat: user api 서비스 코드 작성

* feat: 연동에 필요한 모델 제작

* feat: 로그인 api 연동

* refactor: 파일 이름 오타 수정

* chore: 필요없는 model 제거

* refactor: 서버 값 반영해 수정

snake case에서 camel Case로 json model을 수정했습니다.

* feat: 회원가입 api 연동

* feat: 내 정보 조회 및 로그아웃 api 연동

* fix: 에러 메시지 출력 제거

* fix: 페이지 이동 경로 변수명으로 변경

* refactor: trim()을 view에서 처리하도록 이동

* refactor: 함수 이름 toggle로 변경

* refactor: 위젯에 const 추가

* refactor: SizedBox 위젯 변수로 수정

* refactor: 로그인 메서드 이름 변경

* refactor: StateNotifier -> Notifier로 수정

* refactor: TokenStorage를 provider로 분리

* refactor: SizedBox 변수 수정

+++ lib/presentation/main/view/post_detail_view.dart
@@ -7,6 +7,7 @@ import 'package:falletter_mobile_v2/core/constants/color.dart';
 import 'package:falletter_mobile_v2/core/constants/text_style.dart';
 import 'package:falletter_mobile_v2/core/router/route_paths.dart';
 import 'package:falletter_mobile_v2/core/utils/time_utils.dart';
+import 'package:falletter_mobile_v2/presentation/mypage/provider/user_info_provider.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
 import 'package:go_router/go_router.dart';
@@ -31,6 +32,7 @@ class _PostDetailViewState extends ConsumerState<PostDetailView> {
   void initState() {
     _commentController.addListener(_onChanged);
     Future.microtask(() {
+      ref.read(userInfoProvider.notifier).getUserInfo();
       ref.read(postsDetailProvider.notifier).loadDetailPost(widget.postId);
     });
     super.initState();
@@ -51,6 +53,7 @@ class _PostDetailViewState extends ConsumerState<PostDetailView> {
   @override
   Widget build(BuildContext context) {
     final post = ref.watch(postsDetailProvider);
+    final myInfo = ref.watch(userInfoProvider);
 
     if (post == null) {
       return Container(
@@ -205,7 +208,7 @@ class _PostDetailViewState extends ConsumerState<PostDetailView> {
                                 Row(
                                   children: [
                                     Text(
-                                      comment.userName, //TODO: 익명 닉네임으로 수정하기 (아직 서버에서 수정 안 됨)
+                                      comment.anonymousNickname,
                                       style: commentInfoStyle,
                                     ),
                                     SizedBox(width: 8),
@@ -223,18 +226,19 @@ class _PostDetailViewState extends ConsumerState<PostDetailView> {
                               ],
                             ),
                           ),
-                          IconButton( //TODO: 내 정보 조회 연동되면 내 아이디랑 댓글 아이디 같을 때만 버튼 표시하도록 수정
-                            onPressed: () {
-                              ref.read(postsDetailProvider.notifier).deleteComment(
-                                  comment.commentId,
-                                  post.id
-                              );
-                            },
-                            icon: Icon(
-                              Symbols.delete,
-                              color: FalletterColor.gray400,
+                          if (myInfo.value?.id != null && myInfo.value?.id == comment.userId)
+                            IconButton(
+                              onPressed: () {
+                                ref.read(postsDetailProvider.notifier).deleteComment(
+                                    comment.commentId,
+                                    post.id
+                                );
+                              },
+                              icon: Icon(
+                                Symbols.delete,
+                                color: FalletterColor.gray400,
+                              ),
                             ),
-                          ),
                         ],
                       ),
                     ),
